Refactor for more PIX tests
